+++ R/yaml.R
@@ -34,12 +34,13 @@ yaml_load = function(
 ) {
   if (use_yaml) return(handle_error(
     yaml::yaml.load(x, eval.expr = FALSE, handlers = yaml_handlers(handlers, envir), ...),
-    function(e, loc) {
-      s = e$message
+    function(loc) {
+      s = geterrmessage()
       r = 'line (\\d+), column (\\d+)'
       m = regmatches(s, regexec(r, s, perl = TRUE))[[1]]
       if (length(m) < 3) return()
       m = as.integer(m[-1])  # c(row, col)
+      if (loc != '') loc = paste(' at lines', loc)
       c(
         sprintf('Failed to parse YAML%s:', loc), '',
         append(x, paste0(strrep(' ', m[2]), '^~~~~~'), m[1]), ''
@@ -117,19 +118,27 @@ yaml_handlers = function(h, envir) {
 #' @param x A character vector of the document content.
 #' @param ... Arguments to be passed to `yaml_load()`.
 #' @export
-#' @return A list of components `yaml` and `body`.
+#' @return A list of components `yaml` (the parsed YAML data), `lines` (starting
+#'   and ending line numbers of YAML), and `body` (a character vector of the
+#'   body text). If YAML metadata does not exist in the document, the components
+#'   `yaml` and `lines` will be missing.
 #' @examples
 #' xfun::yaml_body(c('---', 'title: Hello', 'output: markdown::html_document', '---', '', 'Content.'))
 yaml_body = function(x, ...) {
-  i = grep('^---\\s*$', x)
   n = length(x)
-  res = if (n < 2 || length(i) < 2 || (i[1] > 1 && !all(is_blank(x[seq(i[1] - 1)])))) {
-    list(yaml = list(), body = x)
+  res = if (length(i <- locate_yaml(x)) == 0) {
+    list(body = x)
   } else list(
-    yaml = x[i[1]:i[2]], body = c(rep('', i[2]), tail(x, n - i[2]))
+    yaml = x[i[1]:i[2]], body = c(rep('', i[2]), tail(x, n - i[2])), lines = i
   )
-  if ((n <- length(res$yaml)) >= 3) {
-    res$yaml = yaml_load(res$yaml[-c(1, n)], ...)
+  if ((n <- length(res$yaml)) >= 2) {
+    res['yaml'] = list(yaml_load(res$yaml[-c(1, n)], ...))
   }
   res
+}
+
+# find lines of YAML
+locate_yaml = function(x) {
+  i = grep('^---\\s*$', x)
+  if (length(i) > 1 && all(is_blank(x[seq_len(i[1] - 1)]))) i[1:2]
 }